ユーザー削除機能

+++ app/Http/Controllers/UserController.php
@@ -38,11 +38,13 @@ class UserController extends Controller
         abort(404);
     }
 
-    public function destroy()
+    public function destroy($id)
     {
-        $user = User::find(Auth::id());
-        $user->delete();
-
-        return redirect('/');
+        if ($id == Auth::id()) {
+            $user = User::findOrFail($id);
+            $user->delete();
+            return redirect('logout');
+        }
+        abort(404);
     }
 }
